@@ -13,18 +13,17 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/core": "0.6.1",
-    "@angular/animations": "^7.2.13",
+    "@angular/animations": "^8.2.6",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "7.2.13",
-    "@angular/compiler": "7.2.13",
-    "@angular/core": "7.2.13",
-    "@angular/forms": "7.2.13",
-    "@angular/http": "7.2.13",
+    "@angular/common": "8.2.6",
+    "@angular/compiler": "8.2.6",
+    "@angular/core": "8.2.6",
+    "@angular/forms": "8.2.6",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "7.2.13",
-    "@angular/platform-browser-dynamic": "7.2.13",
-    "@angular/router": "7.2.13",
-    "@angular/service-worker": "7.2.13",
+    "@angular/platform-browser": "8.2.6",
+    "@angular/platform-browser-dynamic": "8.2.6",
+    "@angular/router": "8.2.6",
+    "@angular/service-worker": "8.2.6",
     "@aspnet/signalr": "^1.1.4",
     "@material/button": "^1.1.1",
     "@material/drawer": "^2.0.0",
@@ -41,20 +40,20 @@
     "date-fns": "^1.30.1",
     "lodash-es": "^4.17.11",
     "npm": "^6.9.0",
-    "rxjs": "^6.4.0",
-    "zone.js": "^0.8.29"
+    "rxjs": "^6.5.3",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "node-sass": "^4.12.0",
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "7.3.8",
-    "@angular/compiler-cli": "7.2.13",
-    "@angular/language-service": "7.2.13",
+    "@angular-devkit/build-angular": "~0.803.4",
+    "@angular/cli": "8.3.4",
+    "@angular/compiler-cli": "8.2.6",
+    "@angular/language-service": "8.2.6",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/lodash-es": "^4.17.1",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^3.1.4",
@@ -66,6 +65,6 @@
     "protractor": "^5.4.2",
     "ts-node": "~3.2.0",
     "tslint": "^5.12.0",
-    "typescript": "3.2.4"
+    "typescript": "3.5.3"
   }
 }
